+++ Buildings/Resources/src/ThermalZones/EnergyPlus/C-Sources/OutputVariableAllocate.c
@@ -14,7 +14,7 @@
 FMUOutputVariable* checkForDoubleOutputVariableDeclaration(
   const struct FMUBuilding* bui,
   const char* fmiName){
-  int iComVar;
+  size_t iComVar;
 
   for(iComVar = 0; iComVar < bui->nOutputVariables; iComVar++){
     FMUOutputVariable* ptrOutVar = (FMUOutputVariable*)(bui->outputVariables[iComVar]);
@@ -62,7 +62,7 @@ void* EnergyPlusOutputVariableAllocate(
   const char* fmuName,
   const char* buildingsLibraryRoot,
   const double initialTime,
-  const int logLevel,
+  const size_t logLevel,
   int printUnit,
   void (*SpawnMessage)(const char *string),
   void (*SpawnError)(const char *string),
